backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
distro==1.9.0
fastapi==0.104.1
frozenlist==1.8.0
google-ai-generativelanguage==0.4.0
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.3.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
multidict==6.7.0
openai==1.3.0
packaging==25.0
pdf2image==1.16.3
Pillow==10.1.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyMuPDF==1.26.5
pyparsing==3.2.5
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.6
requests==2.32.5
rsa==4.9.1
ruff==0.14.2
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
yarl==1.22.0
